+++ CHANGELOG.md
@@ -22,8 +22,12 @@ Types of changes:
 ### Added
 
 - Enable multi-metamap builder to make replacements to the source posix path.
-- All cui file to do transformations
+- All cui file to do transformations.
 - Script to transform file from UTF-8 (or other encoding) to ASCII.
+
+### Fixed
+
+- Posix can't have quotes in the classpath when supplying argument to MM
 
 
 ## [0.3.4] - 2023-06-07